+++ controlplane/src/core/repositories/GraphCompositionRepository.ts
@@ -1,11 +1,11 @@
 import { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
-import { and, count, desc, eq, gt, lt } from 'drizzle-orm';
+import { SQL, and, count, desc, eq, gt, lt } from 'drizzle-orm';
 import { JsonValue } from '@bufbuild/protobuf';
 import { FastifyBaseLogger } from 'fastify';
 import { splitLabel } from '@wundergraph/cosmo-shared';
 import * as schema from '../../db/schema.js';
 import { graphCompositions, graphCompositionSubgraphs, schemaVersion, targets, users } from '../../db/schema.js';
-import { DateRange, GraphCompositionDTO, SubgraphDTO } from '../../types/index.js';
+import { DateRange, GraphCompositionDTO } from '../../types/index.js';
 import { FederatedGraphRepository } from './FederatedGraphRepository.js';
 
 export class GraphCompositionRepository {
@@ -17,34 +17,34 @@ export class GraphCompositionRepository {
   public async addComposition({
     fedGraphSchemaVersionId,
     compositionErrorString,
-    routerConfig,
     routerConfigSignature,
     subgraphSchemaVersionIds,
     composedBy,
     admissionErrorString,
     deploymentErrorString,
+    isFeatureFlagComposition,
   }: {
     fedGraphSchemaVersionId: string;
     compositionErrorString: string;
-    routerConfig?: JsonValue;
     routerConfigSignature?: string;
     subgraphSchemaVersionIds: string[];
     composedBy: string;
     admissionErrorString?: string;
     deploymentErrorString?: string;
+    isFeatureFlagComposition: boolean;
   }) {
     await this.db.transaction(async (tx) => {
       const insertedComposition = await tx
         .insert(graphCompositions)
         .values({
           schemaVersionId: fedGraphSchemaVersionId,
-          routerConfig: routerConfig || null,
           compositionErrors: compositionErrorString,
           isComposable: compositionErrorString === '',
           routerConfigSignature,
           createdBy: composedBy,
           deploymentError: deploymentErrorString,
           admissionError: admissionErrorString,
+          isFeatureFlagComposition,
         })
         .returning()
         .execute();
@@ -60,6 +60,27 @@ export class GraphCompositionRepository {
           .execute();
       }
     });
+  }
+
+  public updateComposition({
+    fedGraphSchemaVersionId,
+    admissionErrorString,
+    deploymentErrorString,
+    routerConfigSignature,
+  }: {
+    fedGraphSchemaVersionId: string;
+    admissionErrorString?: string;
+    deploymentErrorString?: string;
+    routerConfigSignature?: string;
+  }) {
+    return this.db
+      .update(graphCompositions)
+      .set({
+        deploymentError: deploymentErrorString,
+        admissionError: admissionErrorString,
+        routerConfigSignature,
+      })
+      .where(eq(graphCompositions.schemaVersionId, fedGraphSchemaVersionId));
   }
 
   public async getGraphComposition(input: {
@@ -181,6 +202,7 @@ export class GraphCompositionRepository {
         lastUpdatedAt: graphCompositionSubgraphs.createdAt,
         websocketSubprotocol: schema.subgraphs.websocketSubprotocol,
         isEventDrivenGraph: schema.subgraphs.isEventDrivenGraph,
+        isFeatureSubgraph: schema.subgraphs.isFeatureSubgraph,
       })
       .from(graphCompositionSubgraphs)
       .innerJoin(graphCompositions, eq(graphCompositions.id, graphCompositionSubgraphs.graphCompositionId))
@@ -206,14 +228,25 @@ export class GraphCompositionRepository {
     limit,
     offset,
     dateRange,
+    excludeFeatureFlagCompositions,
   }: {
     fedGraphTargetId: string;
     organizationId: string;
     limit: number;
     offset: number;
     dateRange: DateRange;
+    excludeFeatureFlagCompositions: boolean;
   }): Promise<GraphCompositionDTO[]> {
     const fedRepo = new FederatedGraphRepository(this.logger, this.db, organizationId);
+    const conditions: SQL<unknown>[] = [
+      eq(schemaVersion.targetId, fedGraphTargetId),
+      gt(graphCompositions.createdAt, new Date(dateRange.start)),
+      lt(graphCompositions.createdAt, new Date(dateRange.end)),
+    ];
+
+    if (excludeFeatureFlagCompositions) {
+      conditions.push(eq(graphCompositions.isFeatureFlagComposition, false));
+    }
 
     const resp = await this.db
       .select({
@@ -230,13 +263,7 @@ export class GraphCompositionRepository {
       .from(graphCompositions)
       .innerJoin(schemaVersion, eq(schemaVersion.id, graphCompositions.schemaVersionId))
       .leftJoin(users, eq(graphCompositions.createdBy, users.id))
-      .where(
-        and(
-          eq(schemaVersion.targetId, fedGraphTargetId),
-          gt(graphCompositions.createdAt, new Date(dateRange.start)),
-          lt(graphCompositions.createdAt, new Date(dateRange.end)),
-        ),
-      )
+      .where(and(...conditions))
       .orderBy(desc(schemaVersion.createdAt))
       .limit(limit)
       .offset(offset)
@@ -267,10 +294,22 @@ export class GraphCompositionRepository {
   public async getGraphCompositionsCount({
     fedGraphTargetId,
     dateRange,
+    excludeFeatureFlagCompositions,
   }: {
     fedGraphTargetId: string;
     dateRange: DateRange;
+    excludeFeatureFlagCompositions: boolean;
   }): Promise<number> {
+    const conditions: SQL<unknown>[] = [
+      eq(schemaVersion.targetId, fedGraphTargetId),
+      gt(graphCompositions.createdAt, new Date(dateRange.start)),
+      lt(graphCompositions.createdAt, new Date(dateRange.end)),
+    ];
+
+    if (excludeFeatureFlagCompositions) {
+      conditions.push(eq(graphCompositions.isFeatureFlagComposition, false));
+    }
+
     const compositionsCount = await this.db
       .select({
         count: count(),
@@ -278,13 +317,7 @@ export class GraphCompositionRepository {
       .from(graphCompositions)
       .innerJoin(schemaVersion, eq(schemaVersion.id, graphCompositions.schemaVersionId))
       .leftJoin(users, eq(graphCompositions.createdBy, users.id))
-      .where(
-        and(
-          eq(schemaVersion.targetId, fedGraphTargetId),
-          gt(graphCompositions.createdAt, new Date(dateRange.start)),
-          lt(graphCompositions.createdAt, new Date(dateRange.end)),
-        ),
-      )
+      .where(and(...conditions))
       .execute();
 
     if (compositionsCount.length === 0) {