+++ router-tests/update-config.sh
@@ -4,6 +4,7 @@
 
 ## using source code
 
+echo "Generating config using 'wgc router compose'"
 cd "../cli" || exit
 pnpm wgc router compose -i ../demo/graph.yaml -o ../router-tests/testenv/testdata/config.json
 
@@ -13,9 +14,12 @@ pnpm wgc router compose -i ../demo/graph.yaml -o ../router-tests/testenv/testdat
 
 # go composition wrapper
 
-go run ../demo/cmd/generateconfig/main.go
+# Can't use go wrapper because feature-graphs is not implemented in go yet
+#echo "Generating config using go wrapper"
+#go run ../demo/cmd/generateconfig/main.go
 
 # format test config
 
+echo "Formatting config"
 jq . ../router-tests/testenv/testdata/config.json > ../router-tests/testenv/testdata/config.json.tmp
 mv ../router-tests/testenv/testdata/config.json.tmp ../router-tests/testenv/testdata/config.json