+++ router/core/context.go
@@ -2,12 +2,12 @@ package core
 
 import (
 	"context"
+	"github.com/wundergraph/graphql-go-tools/v2/pkg/engine/datasource/httpclient"
+	"go.opentelemetry.io/otel/attribute"
 	"net/http"
 	"net/url"
 	"sync"
 	"time"
-
-	"github.com/wundergraph/graphql-go-tools/v2/pkg/engine/datasource/httpclient"
 
 	"github.com/wundergraph/cosmo/router/pkg/authentication"
 	ctrace "github.com/wundergraph/cosmo/router/pkg/trace"
@@ -20,6 +20,7 @@ type key string
 
 const requestContextKey = key("request")
 const subgraphsContextKey = key("subgraphs")
+const baseAttributesContextKey = key("baseOtelAttributes")
 
 var _ RequestContext = (*requestContext)(nil)
 
@@ -447,6 +448,15 @@ func withSubgraphs(ctx context.Context, subgraphs []Subgraph) context.Context {
 func subgraphsFromContext(ctx context.Context) []Subgraph {
 	subgraphs, _ := ctx.Value(subgraphsContextKey).([]Subgraph)
 	return subgraphs
+}
+
+func withBaseAttributes(ctx context.Context, attributes []attribute.KeyValue) context.Context {
+	return context.WithValue(ctx, baseAttributesContextKey, attributes)
+}
+
+func baseAttributesFromContext(ctx context.Context) []attribute.KeyValue {
+	attributes, _ := ctx.Value(baseAttributesContextKey).([]attribute.KeyValue)
+	return attributes
 }
 
 func buildRequestContext(w http.ResponseWriter, r *http.Request, opContext *operationContext, requestLogger *zap.Logger) *requestContext {