+++ controlplane/CHANGELOG.md
@@ -4,6 +4,18 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.92.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.92.0...controlplane@0.92.1) (2024-07-03)
+
+### Bug Fixes
+
+* limit check and remove router_config_path ([#911](https://github.com/wundergraph/cosmo/issues/911)) ([93180ed](https://github.com/wundergraph/cosmo/commit/93180edbe3f2f84e1c95f3d3a9acd57ec337a6dc)) (@JivusAyrus)
+
+# [0.92.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.91.5...controlplane@0.92.0) (2024-07-03)
+
+### Features
+
+* feature flags ([#853](https://github.com/wundergraph/cosmo/issues/853)) ([5461bb5](https://github.com/wundergraph/cosmo/commit/5461bb5a529decd51a1b22be0a5301936b8ad392)) (@JivusAyrus)
+
 ## [0.91.5](https://github.com/wundergraph/cosmo/compare/controlplane@0.91.4...controlplane@0.91.5) (2024-07-02)
 
 ### Bug Fixes